% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AND.R
\name{AND}
\alias{AND}
\title{Relational-operator-function to construct a custom search statement}
\usage{
AND(..., negate = FALSE)
}
\arguments{
\item{...}{a combination of criteria constructor functions with its arguments.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{request} parameter in
 \code{ImapCon$search()} function.
}
\description{
Relational-operator-function to construct a custom search statement
}
\examples{
\dontrun{
# select folder & search
con$select_folder(name = "INBOX")
# search for messages SINCE "30-Aug-2019" AND SMALLER than 512KB.
res <- con$search(request = AND(sent_since(date_char = "30-Aug-2019"),
                                smaller_than(size = 512000)))
}
}
\seealso{
Other custom search:
\code{\link{ImapCon}},
\code{\link[=OR]{OR()}},
\code{\link[=before]{before()}},
\code{\link[=flag]{flag()}},
\code{\link[=larger_than]{larger_than()}},
\code{\link[=older_than]{older_than()}},
\code{\link[=on]{on()}},
\code{\link[=sent_before]{sent_before()}},
\code{\link[=sent_on]{sent_on()}},
\code{\link[=sent_since]{sent_since()}},
\code{\link[=since]{since()}},
\code{\link[=smaller_than]{smaller_than()}},
\code{\link[=string]{string()}},
\code{\link[=younger_than]{younger_than()}}
}
\concept{custom search}
