% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-copyMsg.R
\name{check_args_copyMsg}
\alias{check_args_copyMsg}
\title{Copy Message Arguments Check}
\usage{
check_args_copyMsg(imapconf, msg_id, by, to_mbox, reselect_mbox,
  logical_output, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{to_mbox}{A character string specifying the mailbox to which messages
will be copied to.}

\item{reselect_mbox}{If \code{TRUE}, calls \code{selectMailbox(mbox = to_mbox)}
before returning the output. Default is \code{FALSE} for moving and
copying operations, whereas it is \code{TRUE} for renaming mailboxes.}

\item{logical_output}{A logical. If \code{TRUE}, simplifies the output
returning \code{TRUE} to indicate the success of the mentioned operation.
If \code{FALSE}, returns a list of length \code{2} containing the
\code{imapconf} object (IMAP settings) and the previously informed
message ids (parameter \code{msg_id}), so they can be passed as
arguments to another function. Default is \code{FALSE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking the arguments
    used in \link{copyMsg} function.
}
\seealso{
Other miscellaneous helper: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_getMaxMinId}}

Other check args: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_customSearch}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgMetadata}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{check_args_getMaxMinId}},
  \code{\link{check_args_moveMsg}},
  \code{\link{check_args_removeFlags}},
  \code{\link{check_args_searchDate}},
  \code{\link{check_args_searchFlag}},
  \code{\link{check_args_searchPeriod}},
  \code{\link{check_args_searchSize}},
  \code{\link{check_args_searchString}},
  \code{\link{check_args_searchWithin}}
}
\concept{check args}
\concept{miscellaneous helper}
\keyword{internal}
