% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-date.R
\name{check_args_date}
\alias{check_args_date}
\title{Custom Search Specific Check}
\usage{
check_args_date(date_char, negate)
}
\arguments{
\item{date_char}{A character vector with format "DD-Mon-YYYY",
e.g. "01-Apr-2019". We opted not to use objects of type "date", since IMAP
servers like this not so common date format.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking specific arguments
    used in \link{custom_search}.
}
\seealso{
Other customsearch checkargs functions: \code{\link{check_args_flag}},
  \code{\link{check_args_size}},
  \code{\link{check_args_string}},
  \code{\link{check_args_within}}
}
\concept{customsearch checkargs functions}
\keyword{internal}
