% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-expunge.R
\name{check_args_expunge}
\alias{check_args_expunge}
\title{Expunge Message Arguments Check}
\usage{
check_args_expunge(imapconf, specific_UID, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{specific_UID}{A numeric vector containing one or more messages UIDs, if
specific messages should be expunged and not the whole mailbox.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking the arguments
    used in \link{expunge} function.
}
\seealso{
Other miscellaneous helper: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_copyMsg}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_getMaxMinId}}

Other check args: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_copyMsg}},
  \code{\link{check_args_customSearch}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgMetadata}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{check_args_getMaxMinId}},
  \code{\link{check_args_moveMsg}},
  \code{\link{check_args_removeFlags}},
  \code{\link{check_args_searchDate}},
  \code{\link{check_args_searchFlag}},
  \code{\link{check_args_searchPeriod}},
  \code{\link{check_args_searchSize}},
  \code{\link{check_args_searchString}},
  \code{\link{check_args_searchWithin}}
}
\concept{check args}
\concept{miscellaneous helper}
\keyword{internal}
