% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-fetch-msg-header.R
\name{check_args_fetch_msg_header}
\alias{check_args_fetch_msg_header}
\title{Header Fetch Arguments Check}
\usage{
check_args_fetch_msg_header(imapconf, msg_id, by, fields, negate_fields,
  peek, partial, write_to_disk, keep_in_mem, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configure_imap} and with a \code{mbox} item added after a
mailbox selection with \link{select_mailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{fields}{A character vector specifying which field(s) to fetch from
messages. See \link{section_or_field_options}}

\item{negate_fields}{If \code{TRUE}, negates the search and seeks for
"NOT in the field". Default is \code{FALSE}.}

\item{peek}{If \code{TRUE}, it does not mark messages as "read" after
fetching. Default is \code{TRUE}.}

\item{partial}{\code{NULL} or a character string with format
"startchar.endchar" indicating the size (in characters) of a message slice
to fetch. Default is \code{NULL}, which fetchs the full specified content.}

\item{write_to_disk}{If \code{TRUE} writes fetch content of each message
to the disk as a text file in the working directory. Default is \code{FALSE}.}

\item{keep_in_mem}{If \code{TRUE} keeps a copy of fetch results as an
list in the R session when \code{write_to_disk = TRUE}. Default is
\code{FALSE}. It can only be set \code{TRUE} when
\code{write_to_disk = TRUE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking arguments used in
    fetch functions.
}
\seealso{
Other fetch helper: \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{clean_fetch_results}},
  \code{\link{decode_base64_text_if_needed}}

Other check args: \code{\link{check_args_add_replace_flags}},
  \code{\link{check_args_copy_msg}},
  \code{\link{check_args_custom_search}},
  \code{\link{check_args_delete_msg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{check_args_get_max_min_id}},
  \code{\link{check_args_move_msg}},
  \code{\link{check_args_remove_flags}},
  \code{\link{check_args_search_date}},
  \code{\link{check_args_search_flag}},
  \code{\link{check_args_search_period}},
  \code{\link{check_args_search_size}},
  \code{\link{check_args_search_string}},
  \code{\link{check_args_search_within}}
}
\concept{check args}
\concept{fetch helper}
\keyword{internal}
