% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-fetchMsgMetadata.R
\name{check_args_fetchMsgMetadata}
\alias{check_args_fetchMsgMetadata}
\title{Metadata Fetch Arguments Check}
\usage{
check_args_fetchMsgMetadata(imapconf, msg_id, by, metadata, write_to_file,
  keep_in_mem, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{metadata}{A character vector or string specifying one or more items of
the metadata of a message to fetch. See \link{metadata_options}.}

\item{write_to_file}{If \code{TRUE} writes fetch content of each message
to the disk as a text file in the working directory. Default is \code{FALSE}.}

\item{keep_in_mem}{If \code{TRUE} keeps a copy of fetch results as an
list in the R session when \code{write_to_file = TRUE}. Default is
\code{FALSE}. It can only be set \code{TRUE} when
\code{write_to_file = TRUE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking arguments used in
    fetch functions.
}
\seealso{
Other fetch helper: \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{clean_fetch_results}},
  \code{\link{decode_base64_text_if_TRUE}}

Other check args: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_copyMsg}},
  \code{\link{check_args_customSearch}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{check_args_getMaxMinId}},
  \code{\link{check_args_moveMsg}},
  \code{\link{check_args_removeFlags}},
  \code{\link{check_args_searchDate}},
  \code{\link{check_args_searchFlag}},
  \code{\link{check_args_searchPeriod}},
  \code{\link{check_args_searchSize}},
  \code{\link{check_args_searchString}},
  \code{\link{check_args_searchWithin}}
}
\concept{check args}
\concept{fetch helper}
\keyword{internal}
