% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-flag.R
\name{check_args_flag}
\alias{check_args_flag}
\title{Custom Search Specific Check}
\usage{
check_args_flag(flag, negate)
}
\arguments{
\item{flag}{A string specifying the flag to be used for filtering messages.
Use \link{flag_options} to list the flags available in your IMAP server.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking specific arguments
    used in \link{custom_search}.
}
\seealso{
Other customsearch checkargs functions: \code{\link{check_args_date}},
  \code{\link{check_args_size}},
  \code{\link{check_args_string}},
  \code{\link{check_args_within}}
}
\concept{customsearch checkargs functions}
\keyword{internal}
