% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-getMaxMinId.R
\name{check_args_getMaxMinId}
\alias{check_args_getMaxMinId}
\title{Max/Min Message Arguments Check}
\usage{
check_args_getMaxMinId(imapconf, by, flag, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flag}{A string specifying the flag to be used for filtering messages.
Use \link{flag_options} to list the common flags used by IMAP
servers. Default is \code{NULL}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking the arguments
    used in \link{getMaxId} and \link{getMinId}  functions.
}
\seealso{
Other miscellaneous helper: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_copyMsg}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_expunge}}

Other check args: \code{\link{check_args_addReplaceFlags}},
  \code{\link{check_args_copyMsg}},
  \code{\link{check_args_customSearch}},
  \code{\link{check_args_deleteMsg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgMetadata}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{check_args_moveMsg}},
  \code{\link{check_args_removeFlags}},
  \code{\link{check_args_searchDate}},
  \code{\link{check_args_searchFlag}},
  \code{\link{check_args_searchPeriod}},
  \code{\link{check_args_searchSize}},
  \code{\link{check_args_searchString}},
  \code{\link{check_args_searchWithin}}
}
\concept{check args}
\concept{miscellaneous helper}
\keyword{internal}
