% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-list-attachments.R
\name{check_args_list_attachments}
\alias{check_args_list_attachments}
\title{List Attachments Arguments Check}
\usage{
check_args_list_attachments(msg_list)
}
\arguments{
\item{msg_list}{A \code{list} containing the full messages fetched from
the server after using \code{\link{fetch_full_msg}}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for listing attachments
}
\seealso{
Other attachments helper: \code{\link{check_args_get_attachments}},
  \code{\link{has_attachment}}

Other check args: \code{\link{check_args_add_replace_flags}},
  \code{\link{check_args_copy_msg}},
  \code{\link{check_args_custom_search}},
  \code{\link{check_args_delete_msg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_header}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{check_args_get_attachments}},
  \code{\link{check_args_get_max_min_id}},
  \code{\link{check_args_move_msg}},
  \code{\link{check_args_remove_flags}},
  \code{\link{check_args_search_date}},
  \code{\link{check_args_search_flag}},
  \code{\link{check_args_search_period}},
  \code{\link{check_args_search_size}},
  \code{\link{check_args_search_string}},
  \code{\link{check_args_search_within}}
}
\concept{attachments helper}
\concept{check args}
\keyword{internal}
