% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-remove-flags.R
\name{check_args_remove_flags}
\alias{check_args_remove_flags}
\title{Remove Flags Arguments Check}
\usage{
check_args_remove_flags(imapconf, msg_id, by, flags_to_unset, retries)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configure_imap} and with a \code{mbox} item added after a
mailbox selection with \link{select_mailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flags_to_unset}{A character vector containing one ore more flag names to
remove from the specified message ids.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking the arguments
    used in \link{remove_flags} function.
}
\seealso{
Other miscellaneous helpers: \code{\link{check_args_move_msg}}

Other check args: \code{\link{check_args_add_replace_flags}},
  \code{\link{check_args_copy_msg}},
  \code{\link{check_args_custom_search}},
  \code{\link{check_args_delete_msg}},
  \code{\link{check_args_expunge}},
  \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_header}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{check_args_get_attachments}},
  \code{\link{check_args_get_max_min_id}},
  \code{\link{check_args_move_msg}},
  \code{\link{check_args_search_date}},
  \code{\link{check_args_search_flag}},
  \code{\link{check_args_search_period}},
  \code{\link{check_args_search_size}},
  \code{\link{check_args_search_string}},
  \code{\link{check_args_search_within}}
}
\concept{check args}
\concept{miscellaneous helpers}
\keyword{internal}
