% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-size.R
\name{check_args_size}
\alias{check_args_size}
\title{Custom Search Specific Check}
\usage{
check_args_size(size, negate)
}
\arguments{
\item{size}{An integer specifying the size (in number of characters) of
message to be used as search criterion.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking specific arguments
    used in \link{custom_search}.
}
\seealso{
Other customsearch checkargs functions: \code{\link{check_args_date}},
  \code{\link{check_args_flag}},
  \code{\link{check_args_string}},
  \code{\link{check_args_within}}
}
\concept{customsearch checkargs functions}
\keyword{internal}
