% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args-string.R
\name{check_args_string}
\alias{check_args_string}
\title{Custom Search Specific Check}
\usage{
check_args_string(string, section_or_field, negate)
}
\arguments{
\item{string}{A character string specifying the word or expression to
search for in messages.}

\item{section_or_field}{A mandatory character string specifying in which
messages's Section or Header Field to search for the provided string.
For some available options, see \link{section_or_field_options}.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
\code{NULL} if arguments are correct.
}
\description{
Internal helper function for checking specific arguments
    used in \link{custom_search}.
}
\seealso{
Other customsearch checkargs functions: \code{\link{check_args_date}},
  \code{\link{check_args_flag}},
  \code{\link{check_args_size}},
  \code{\link{check_args_within}}
}
\concept{customsearch checkargs functions}
\keyword{internal}
