% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-fetch-results.R
\name{clean_fetch_results}
\alias{clean_fetch_results}
\title{Messages Results Cleaning}
\usage{
clean_fetch_results(msg)
}
\arguments{
\item{msg}{A character string with a message content.}
}
\value{
A string with a cleaned message.
}
\description{
Internal helper function for cleaning a message content in a
    string. It is used with fetch loop functions.
}
\seealso{
Other fetch helper: \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_header}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{decode_base64_text_if_needed}}

Other cleaning: \code{\link{decode_base64_text_if_needed}}
}
\concept{cleaning}
\concept{fetch helper}
\keyword{internal}
