% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-handle.R
\name{config_handle}
\alias{config_handle}
\title{Curl Handle Configuration}
\usage{
config_handle(imapconf)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\code{configure_imap}.}
}
\value{
A curl handle object.
}
\description{
Internal helper function for configuring curl handles used in
    almost all mRpostman functions.
}
\seealso{
Other config: \code{\link{configure_imap}}
}
\concept{config}
\concept{config helper}
\keyword{internal}
