% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyMsg.R
\name{copyMsg}
\alias{copyMsg}
\title{Copy Messages}
\usage{
copyMsg(imapconf, msg_id, by = "MSN", to_mbox, reselect_mbox = FALSE,
  logical_output = TRUE, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{to_mbox}{A character string specifying the mailbox to which messages
will be copied to.}

\item{reselect_mbox}{If \code{TRUE}, calls \code{selectMailbox(mbox = to_mbox)}
before returning the output. Default is \code{FALSE} for moving and
copying operations, whereas it is \code{TRUE} for renaming mailboxes.}

\item{logical_output}{A logical. If \code{TRUE}, simplifies the output
returning \code{TRUE} to indicate the success of the mentioned operation.
If \code{FALSE}, returns a list of length \code{2} containing the
\code{imapconf} object (IMAP settings) and the previously informed
message ids (parameter \code{msg_id}), so they can be passed as
arguments to another function. Default is \code{FALSE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
Depending on the \code{logical_output} parameter, returns a \code{list}
    of length \code{2} containing the \code{imapconf} object and previously
    informed message ids (parameter \code{msg_id}), or a logical vector of
    length \code{1} indicating the success (\code{TRUE}) of the mentioned
    operation.
}
\description{
Copies messages form a mailbox to another.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# copy search results from "Sent" to "INBOX"
results <- imapconf \%>\%
    selectMailbox(mbox = "Sent") \%>\%
    searchBefore(date_char = "10-may-2012") \%$\% #exposition pipe operator - pass two argg
    copyMsg(imapconf = imapconf, msg_id = msg_id, to_mbox = "INBOX")

}

}
\seealso{
Other miscellaneous: \code{\link{addFlags}},
  \code{\link{countMsgs}}, \code{\link{deleteMsg}},
  \code{\link{expunge}}, \code{\link{getMaxId}},
  \code{\link{getMinId}}, \code{\link{moveMsg}},
  \code{\link{removeFlags}}, \code{\link{replaceFlags}}
}
\concept{miscellaneous}
