% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMsgs.R
\name{countMsgs}
\alias{countMsgs}
\title{Count Number of Messages}
\usage{
countMsgs(imapconf, by = "MSN", flag, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flag}{A string specifying the flag to be used for filtering messages.
Use \link{flag_options} to list the common flags used by IMAP
servers.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A numeric \code{vector} of length \code{1} containing the number of
    messages that meet the specified criteria.
}
\description{
Counts the number of messages with a respective flag in a mailbox.
}
\note{
\code{copyMsg} uses \code{ESEARCH} extension. Check if your server
    supports ESEARCH capability with \link{listServerCapabilities}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

results <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    countMsgs(by = "UID", flag = "SEEN")

}

}
\seealso{
Other miscellaneous: \code{\link{addFlags}},
  \code{\link{copyMsg}}, \code{\link{deleteMsg}},
  \code{\link{expunge}}, \code{\link{getMaxId}},
  \code{\link{getMinId}}, \code{\link{moveMsg}},
  \code{\link{removeFlags}}, \code{\link{replaceFlags}}
}
\concept{miscellaneous}
