% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-base64-text-if-needed.R
\name{decode_base64_text_if_needed}
\alias{decode_base64_text_if_needed}
\title{Messages Results Decoding}
\usage{
decode_base64_text_if_needed(msg)
}
\arguments{
\item{msg}{A character string with a message content.}
}
\value{
A string with a decoded message, if it was base64 coded.
}
\description{
Internal helper function for decoding base64 message results.
}
\seealso{
Other fetch helper: \code{\link{check_args_fetch_full_msg}},
  \code{\link{check_args_fetch_msg_header}},
  \code{\link{check_args_fetch_msg_metadata}},
  \code{\link{check_args_fetch_msg_text}},
  \code{\link{clean_fetch_results}}

Other cleaning: \code{\link{clean_fetch_results}}
}
\concept{cleaning}
\concept{fetch helper}
\keyword{internal}
