% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-base64-text-if-TRUE.R
\name{decode_base64_text_if_TRUE}
\alias{decode_base64_text_if_TRUE}
\title{Messages Results Decoding}
\usage{
decode_base64_text_if_TRUE(msg)
}
\arguments{
\item{msg}{A character string with a message content.}
}
\value{
A string with a decoded message, if it was base64 coded.
}
\description{
Internal helper function for decoding base64 message results.
}
\seealso{
Other fetch helper: \code{\link{check_args_fetchFullMsg}},
  \code{\link{check_args_fetchMsgHeader}},
  \code{\link{check_args_fetchMsgMetadata}},
  \code{\link{check_args_fetchMsgText}},
  \code{\link{clean_fetch_results}}

Other cleaning: \code{\link{clean_fetch_results}}
}
\concept{cleaning}
\concept{fetch helper}
\keyword{internal}
