% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-searchrequest-period.R
\name{define_searchrequest_period}
\alias{define_searchrequest_period}
\title{Period Search Request}
\usage{
define_searchrequest_period(operation1, since_date_char, operation2,
  before_date_char, negate, by, flag, esearch, handle)
}
\arguments{
\item{operation1}{A character indicating the first argument
(SINCE and variations) to the custom request.}

\item{since_date_char}{A character vector with format "DD-Mon-YYYY",
e.g. "01-Apr-2019" indicating the start date (inclusive). We opted not to
use objects of type "date", since IMAP servers like this not so common date
format.}

\item{operation2}{A character indicating the first argument
(BEFORE and variations) to the custom request.}

\item{before_date_char}{A character vector with format "DD-Mon-YYYY",
e.g. "01-Apr-2019" indicating the stop date (exclusive). We opted not to
use objects of type "date", since IMAP server likes this not so common date
format.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flag}{Optional parameter that adds a flag filter to the search. Use
\link{flag_options} to list the common flags used by IMAP servers.
Default is \code{NULL}.}

\item{esearch}{A logical. Default is \code{FALSE}. If the IMAP server has
\code{ESEARCH} capability, it can be used to optimize search results. It
allows to condense results to message sets to cut down on transmission
costs, e.g. \code{1:5} instead of writing ids individually such as
\code{1,2,3,4,5}. It can be used along with buffersize to avoid results
stripping. Check if your IMAP server supports \code{SEARCH} with
\code{\link{list_server_capabilities}}.}

\item{handle}{A curl handle object.}
}
\value{
A curl handle object containing the custom request.
}
\description{
Internal helper function for defining curl custom request to be
    used with each search function
}
\seealso{
Other search helper: \code{\link{check_args_custom_search}},
  \code{\link{check_args_search_date}},
  \code{\link{check_args_search_flag}},
  \code{\link{check_args_search_period}},
  \code{\link{check_args_search_size}},
  \code{\link{check_args_search_string}},
  \code{\link{check_args_search_within}},
  \code{\link{define_searchrequest_custom}},
  \code{\link{define_searchrequest_date}},
  \code{\link{define_searchrequest_flag}},
  \code{\link{define_searchrequest_size}},
  \code{\link{define_searchrequest_string}},
  \code{\link{define_searchrequest_within}},
  \code{\link{fix_search_stripping}}

Other define searchrequest: \code{\link{define_searchrequest_custom}},
  \code{\link{define_searchrequest_date}},
  \code{\link{define_searchrequest_flag}},
  \code{\link{define_searchrequest_size}},
  \code{\link{define_searchrequest_string}},
  \code{\link{define_searchrequest_within}}
}
\concept{define searchrequest}
\concept{search helper}
\keyword{internal}
