% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteMsg.R
\name{deleteMsg}
\alias{deleteMsg}
\title{Delete Messages}
\usage{
deleteMsg(imapconf, msg_id, by = "MSN", logical_output = TRUE,
  retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{logical_output}{A logical. If \code{TRUE}, simplifies the output
returning \code{TRUE} to indicate the success of the mentioned operation.
If \code{FALSE}, returns a list of length \code{2} containing the
\code{imapconf} object (IMAP settings) and the previously informed
message ids (parameter \code{msg_id}), so they can be passed as
arguments to another function. Default is \code{FALSE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
Depending on the \code{logical_output} parameter, returns a \code{list}
    of length \code{2} containing the \code{imapconf} object and previously
    informed message ids (parameter \code{msg_id}), or a logical vector of
    length \code{1} indicating the success (\code{TRUE}) of the mentioned
    operation.
}
\description{
Deletes messages from a mailbox.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# deleting based on search results
result1 <- imapconf \%>\%
    selectMailbox(mbox = "TAM") \%>\%
    searchBefore(date_char = "10-may-2012", by = "UID") \%$\% #modified pipe operator - pass two argg
    deleteMsg(imapconf = imapconf, msg_id = msg_id)


# deleting a specific msg_id without a previous search
result2 <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    deleteMsg(msg_id = 66128)

}

}
\seealso{
Other miscellaneous: \code{\link{addFlags}},
  \code{\link{copyMsg}}, \code{\link{countMsgs}},
  \code{\link{expunge}}, \code{\link{getMaxId}},
  \code{\link{getMinId}}, \code{\link{moveMsg}},
  \code{\link{removeFlags}}, \code{\link{replaceFlags}}
}
\concept{miscellaneous}
