% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examineMailbox.R
\name{examineMailbox}
\alias{examineMailbox}
\title{Examine Mailbox}
\usage{
examineMailbox(imapconf, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\code{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \code{selectMailbox}}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A \code{vector} containing the (\code{"EXISTS"}) and the
    (\code{"RECENT"}) number of messages in the selected mailbox.
}
\description{
Retrieve the number of recent messages and the total number of
    messages in a mailbox.
}
\examples{
\dontrun{

# configure IMAP
#' library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# list mailboxes
results <- imapconf \%>\%
  listMailboxes() \%>\%
  selectMailbox(mbox = "Sent") \%>\%
  examineMailbox()

}
}
\seealso{
Other mailbox commands: \code{\link{listMailboxes}},
  \code{\link{listServerCapabilities}},
  \code{\link{renameMailbox}}, \code{\link{selectMailbox}}
}
\concept{mailbox commands}
