% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-complementary-operations.R
\name{execute_complementary_operations}
\alias{execute_complementary_operations}
\title{Execution engine for all the complementary commands}
\usage{
execute_complementary_operations(self, url, handle, customrequest, retries)
}
\arguments{
\item{self}{The R6 connection object.}

\item{url}{A string containing the url from the \code{IMAP_conn$imapconf} object.}

\item{handle}{A curl handle object with the custom request already defined.}

\item{customrequest}{A string containing the custom request to the server that will
be added to the curl handle.}

\item{retries}{Number of attempts to connect and execute the command. Default
  is \code{1}.
@noRd}
}
\description{
Execution engine for all the complementary commands
}
