% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expunge.R
\name{expunge}
\alias{expunge}
\title{Expunge Messages}
\usage{
expunge(imapconf, specific_UID = NULL, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{specific_UID}{A numeric vector containing one or more messages UIDs, if
specific messages should be expunged and not the whole mailbox.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A logical vector of length \code{1} indicating the success
    (\code{TRUE}) of the mentioned operation.
}
\description{
Expunge a specific message (using specific UID) or all messages
    marked as "DELETED" in a mailbox.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# deleting a message using specific UID
result1 <- imapconf \%>\%
    selectMailbox(mbox = "TAM") \%>\%
    deleteMsg(msg_id = 71171, by = "UID") \%$\%
    expunge(imapconf = imapconf, specific_UID = msg_id)


# expunge all message smarked as "DELETED" in INBOX
result2 <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    expunge()

}

}
\seealso{
Other miscellaneous: \code{\link{addFlags}},
  \code{\link{copyMsg}}, \code{\link{countMsgs}},
  \code{\link{deleteMsg}}, \code{\link{getMaxId}},
  \code{\link{getMinId}}, \code{\link{moveMsg}},
  \code{\link{removeFlags}}, \code{\link{replaceFlags}}
}
\concept{miscellaneous}
