% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-search-stripping.R
\name{fix_search_stripping}
\alias{fix_search_stripping}
\title{Fixing Stripped Search Results}
\usage{
fix_search_stripping(response)
}
\arguments{
\item{response}{An integer \code{vector} containing message ids from search.}
}
\value{
An integer \code{vector}.
}
\description{
Fixes stripped results from search operations when server may
    have stripped some lines.
}
\seealso{
Other search helper: \code{\link{check_args_custom_search}},
  \code{\link{check_args_search_date}},
  \code{\link{check_args_search_flag}},
  \code{\link{check_args_search_period}},
  \code{\link{check_args_search_size}},
  \code{\link{check_args_search_string}},
  \code{\link{check_args_search_within}},
  \code{\link{define_searchrequest_custom}},
  \code{\link{define_searchrequest_date}},
  \code{\link{define_searchrequest_flag}},
  \code{\link{define_searchrequest_period}},
  \code{\link{define_searchrequest_size}},
  \code{\link{define_searchrequest_string}},
  \code{\link{define_searchrequest_within}}
}
\concept{fix results}
\concept{search helper}
\keyword{internal}
