% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_searchStripping.R
\name{fix_searchStripping}
\alias{fix_searchStripping}
\title{Fixing Stripped Search Results}
\usage{
fix_searchStripping(response)
}
\arguments{
\item{response}{An integer \code{vector} containing message ids from search.}
}
\value{
An integer \code{vector}.
}
\description{
Fixes stripped results from search operations when server may 
    have stripped some lines.
}
\seealso{
Other search helper: \code{\link{check_args_customSearch}},
  \code{\link{check_args_searchDate}},
  \code{\link{check_args_searchFlag}},
  \code{\link{check_args_searchPeriod}},
  \code{\link{check_args_searchSize}},
  \code{\link{check_args_searchString}},
  \code{\link{check_args_searchWithin}},
  \code{\link{define_searchrequest_custom}},
  \code{\link{define_searchrequest_date}},
  \code{\link{define_searchrequest_flag}},
  \code{\link{define_searchrequest_period}},
  \code{\link{define_searchrequest_size}},
  \code{\link{define_searchrequest_string}},
  \code{\link{define_searchrequest_within}}
}
\concept{fix results}
\concept{search helper}
\keyword{internal}
