% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinId.R
\name{getMinId}
\alias{getMinId}
\title{Get Minimum Message ID in a Mailbox}
\usage{
getMinId(imapconf, by = "MSN", flag, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flag}{A string specifying the flag to be used for filtering messages.
Use \link{flag_options} to list the common flags used by IMAP
servers. Default is \code{NULL}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A numeric \code{vector} of length \code{1} indicating the minimum id.
}
\description{
Get the smaller (minimum) message id in a Mailbox, considering
    messages with a specific flag
}
\note{
\code{getMaxId} uses \code{ESEARCH} extension. Check if your server
    supports \code{ESEARCH} capability with \link{listServerCapabilities}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

results <- imapconf \%>\%
    selectMailbox(mbox = "[Gmail]/Trash") \%>\%
    getMinId(flag = "UNSEEN")

}

}
\seealso{
Other miscellaneous: \code{\link{addFlags}},
  \code{\link{copyMsg}}, \code{\link{countMsgs}},
  \code{\link{deleteMsg}}, \code{\link{expunge}},
  \code{\link{getMaxId}}, \code{\link{moveMsg}},
  \code{\link{removeFlags}}, \code{\link{replaceFlags}}
}
\concept{miscellaneous}
