% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-attachment.R
\name{has_attachment}
\alias{has_attachment}
\title{Attachments Check}
\usage{
has_attachment(msg)
}
\arguments{
\item{msg}{An object of class \code{character} containing the whole MIME
message.}
}
\value{
A \code{logical} containing \code{TRUE} if there is one or more
    attachments and \code{FALSE} if there is none.
}
\description{
Internal helper function for attachment checking in messages.
}
\seealso{
Other attachments helper: \code{\link{check_args_get_attachments}}
}
\concept{attachments helper}
\keyword{internal}
