% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larger-than.R
\name{larger_than}
\alias{larger_than}
\title{Criterion constructor function to be combined in a custom search statement}
\usage{
larger_than(size, negate = FALSE)
}
\arguments{
\item{size}{An integer specifying the number of bytes to be used as
search criterion.}

\item{negate}{If \code{TRUE}, negates the search and seeks for "NOT SEARCH
CRITERIA". Default is \code{FALSE}.}
}
\value{
A search criterion of class \code{imap_search}, to be combined
  into a search statement (see \code{Ops.imap_search}).
}
\description{
Criterion constructor function to be combined in a custom search statement
}
\examples{
\dontrun{
# select folder & search
con$select_folder(name = "INBOX")
# search for messages containing the string "XYZ@k-state.edu" in the
#   "FROM" field OR those that are LARGER than 512KB.
res <- con$search(request = OR(string(expr = "XYZ@k-state.edu",
                                      where = "FROM"),
                               larger_than(size = 512000)))
}

}
\seealso{
Other custom search:
\code{\link[=AND]{AND()}},
\code{\link{ImapCon}},
\code{\link[=OR]{OR()}},
\code{\link[=before]{before()}},
\code{\link[=filter_stored]{filter_stored()}},
\code{\link[=flag]{flag()}},
\code{\link[=fuzzy]{fuzzy()}},
\code{\link[=modseq]{modseq()}},
\code{\link[=older_than]{older_than()}},
\code{\link[=on]{on()}},
\code{\link[=saved_before]{saved_before()}},
\code{\link[=sent_before]{sent_before()}},
\code{\link[=sent_on]{sent_on()}},
\code{\link[=sent_since]{sent_since()}},
\code{\link[=since]{since()}},
\code{\link[=smaller_than]{smaller_than()}},
\code{\link[=string]{string()}},
\code{\link[=verbatim]{verbatim()}},
\code{\link[=younger_than]{younger_than()}}
}
\concept{custom search}
