% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMailboxes.R
\name{listMailboxes}
\alias{listMailboxes}
\title{Listing Mailboxes}
\usage{
listMailboxes(imapconf, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A \code{list} containing the Mailboxes (root and children).
}
\description{
List Mailboxes from IMAP server.
}
\examples{
\dontrun{

# configure IMAP
#' library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# list mailboxes
results <- imapconf \%>\%
  listMailboxes()

}
}
\seealso{
Other mailbox commands: \code{\link{examineMailbox}},
  \code{\link{listServerCapabilities}},
  \code{\link{renameMailbox}}, \code{\link{selectMailbox}}
}
\concept{mailbox commands}
