% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listServerCapabilities.R
\name{listServerCapabilities}
\alias{listServerCapabilities}
\title{IMAP Server Capabilities}
\usage{
listServerCapabilities(imapconf, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
A character \code{vector} containing server's IMAP capacbilities.
}
\description{
Lists IMAP server's capabilities.
}
\examples{
\dontrun{

# configure IMAP
#' library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# list server's capabilities
results <- imapconf \%>\%
  listServerCapabilities()

}

}
\seealso{
Other mailbox commands: \code{\link{examineMailbox}},
  \code{\link{listMailboxes}}, \code{\link{renameMailbox}},
  \code{\link{selectMailbox}}
}
\concept{mailbox commands}
