% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-fetchMsgMetadata_v2.R
\name{loop_fetchMsgMetadata}
\alias{loop_fetchMsgMetadata}
\title{Metadata Fetch Loop}
\usage{
loop_fetchMsgMetadata(new_imapconf, msg_id, by, metadata, write_to_disk,
  keep_in_mem, retries, handle)
}
\arguments{
\item{new_imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{metadata}{A character vector or string specifying one or more items of
the metadata of a message to fetch. See \link{metadata_options}.}

\item{keep_in_mem}{If \code{TRUE} keeps a copy of fetch results as an
list in the R session when \code{write_to_file = TRUE}. Default is
\code{FALSE}. It can only be set \code{TRUE} when
\code{write_to_file = TRUE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}

\item{handle}{A curl handle object.}
}
\value{
A \code{list} or/and text files containing the fetch results.
}
\description{
Internal helper function for loop execution used in fetch
    functions.
}
\seealso{
Other fetch helpers: \code{\link{loop_fetchFullMsg}},
  \code{\link{loop_fetchMsgHeader}},
  \code{\link{loop_fetchMsgText}}

Other loop: \code{\link{loop_fetchFullMsg}},
  \code{\link{loop_fetchMsgHeader}},
  \code{\link{loop_fetchMsgText}}
}
\concept{fetch helpers}
\concept{loop}
\keyword{internal}
