% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-fetchMsgText_v2.R
\name{loop_fetchMsgText}
\alias{loop_fetchMsgText}
\title{Metadata Fetch Loop}
\usage{
loop_fetchMsgText(new_imapconf, msg_id, by, peek, partial, write_to_disk,
  keep_in_mem, retries, handle)
}
\arguments{
\item{new_imapconf}{An object of class \code{imapconf} generated with
\link{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{msg_id}{A numeric vetor containing one or more messages ids.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{peek}{If \code{TRUE}, it does not mark messages as "read" after
fetching. Default is \code{TRUE}.}

\item{partial}{\code{NULL} or a character string with format
"startchar.endchar" indicating the size (in characters) of a message slice
to fetch. Default is \code{NULL}, which fetchs the full specified content.}

\item{keep_in_mem}{If \code{TRUE} keeps a copy of fetch results as an
list in the R session when \code{write_to_file = TRUE}. Default is
\code{FALSE}. It can only be set \code{TRUE} when
\code{write_to_file = TRUE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}

\item{handle}{A curl handle object.}
}
\value{
A \code{list} or/and text files containing the fetch results.
}
\description{
Internal helper function for loop execution used in fetch
    functions.
}
\seealso{
Other fetch helpers: \code{\link{loop_fetchFullMsg}},
  \code{\link{loop_fetchMsgHeader}},
  \code{\link{loop_fetchMsgMetadata}}

Other loop: \code{\link{loop_fetchFullMsg}},
  \code{\link{loop_fetchMsgHeader}},
  \code{\link{loop_fetchMsgMetadata}}
}
\concept{fetch helpers}
\concept{loop}
\keyword{internal}
