% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-options.R
\name{metadata_options}
\alias{metadata_options}
\title{Message Metadata Options}
\usage{
metadata_options()
}
\value{
A \code{vector} containing message metadata fields.
}
\description{
List Metadata fields used in messages.
}
\note{
This function lists the message attributes that
    \href{#method-fetch_metadata}{\code{ImapCon$fetch_metadata()}} accepts.
    The last three (\code{PREVIEW}, \code{SAVEDATE}, \code{MODSEQ}) are
    capability-gated extensions, only sent when the server advertises them.
}
\examples{
\dontrun{

library(mRpostman)
metadata_options()

}
}
\references{
Crispin, M., "Internet Message Access Protocol - Version 4rev1",
    RFC 2060, \doi{10.17487/RFC2060}, December 1996,
    \url{https://www.rfc-editor.org/info/rfc2060}.
}
\concept{options}
