% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olderThan.R
\name{olderThan}
\alias{olderThan}
\title{Criteria Helper Functions for Custom Search}
\usage{
olderThan(seconds, negate = FALSE)
}
\arguments{
\item{seconds}{An integer specifying the number of seconds to be used as
search criterion.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Criteria helper functions to be combined inside a operator helper
    function as a custom request in \link{customSearch}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request =
                 OR(String(
                       section_or_field = "from", string = "allanvcq@gmail.com"),
                    olderThan(
                       seconds = 3600)
                   )
                 )
# searches for messages containing the string "allanvcq@gmail.com" in the field "FROM"
    OR messgaes that are "older" than one hour (messages arrived 3more than
    600 seconds ago).

}

}
\seealso{
Other customsearch helper functions: \code{\link{AND}},
  \code{\link{Before}}, \code{\link{Flag}},
  \code{\link{OR}}, \code{\link{Since}},
  \code{\link{String}}, \code{\link{sentBefore}},
  \code{\link{sentOn}}, \code{\link{sentSince}},
  \code{\link{smallerThan}}, \code{\link{youngerThan}}
}
\concept{customsearch helper functions}
