% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameMailbox.R
\name{renameMailbox}
\alias{renameMailbox}
\title{Rename Mailbox}
\usage{
renameMailbox(imapconf, new_name, reselect_mbox = TRUE, retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\code{configureIMAP} and with a \code{mbox} item added after a
mailbox selection with \link{selectMailbox}.}

\item{new_name}{A string containing the new name to be set.}

\item{reselect_mbox}{If \code{TRUE}, calls \code{selectMailbox(mbox = to_mbox)}
before returning the output. Default is \code{FALSE} for moving and
copying operations, whereas it is \code{TRUE} for renaming mailboxes.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
An \code{imaconf} object with the newer selected mailbox added
    to it. If \code{reselect_mbox} is \code{FALSE}, it returns the
    \code{imaconf} object with the former mailbox name.
}
\description{
Rename a mailbox.
}
\examples{
\dontrun{

# configure IMAP
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# rename mailbox "Sent"
imapconf \%>\%
  selectMailbox(mbox = "Sent") \%>\%
  renameMailbox(new_name = "Sent Mail")

}
}
\seealso{
Other mailbox commands: \code{\link{examineMailbox}},
  \code{\link{listMailboxes}},
  \code{\link{listServerCapabilities}},
  \code{\link{selectMailbox}}
}
\concept{base}
\concept{mailbox commands}
