% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfc2047_header_decode.R
\name{rfc2047_header_decode}
\alias{rfc2047_header_decode}
\title{Decode RFC 2047 quoted-printable and base64 headers or strings}
\usage{
rfc2047_header_decode(header)
}
\arguments{
\item{header}{A \code{character} vector containing a string to be decoded.}
}
\value{
A decoded \code{character} vector if applicable.
}
\description{
Decode RFC 2047 quoted-printable and base64 headers or strings
}
\examples{
# Simple quoted-printable string - Portuguese example
qp_encoded <- "Minist=E9rio_da_Educa=E7=E3o"
decoded_string <- rfc2047_header_decode(header = qp_encoded)

# Simple quoted-printable string - French example
qp_encoded <- "Minist=E9rio_da_Educa=E7=E3o"
decoded_string <- rfc2047_header_decode(header = qp_encoded)

# RFC 2047 quoted-printable header - Portuguese example
qp_encoded <- "=?iso-8859-1?Q?DIDEC_Capacita=E7=E3o?="
decoded_string <- rfc2047_header_decode(header = qp_encoded)

# RFC 2047 quoted-printable - German example
qp_encoded <- "=?UTF-8?Q?stern=2Ede_-_t=C3=A4glich?="
decoded_string <- rfc2047_header_decode(header = qp_encoded)

# RFC 2047 base64 - Portuguese example
b64_encoded <- "=?utf-8?B?Sk9BTkEgRlVTQ08gTE9CTyBubyBUZWFtcw==?="
decoded_string <- rfc2047_header_decode(header = b64_encoded)

}
