% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-since.R
\name{search_since}
\alias{search_since}
\title{Search By Internal Date}
\usage{
search_since(imapconf, date_char, negate = FALSE, by = "MSN",
  flag = NULL, esearch = FALSE, return_imapconf = TRUE,
  retries = 2)
}
\arguments{
\item{imapconf}{An object of class \code{imapconf} generated with
\link{configure_imap} and with a \code{mbox} item added after a
mailbox selection with \link{select_mailbox}.}

\item{date_char}{A character vector with format "DD-Mon-YYYY",
e.g. "01-Apr-2019". We opted not to use objects of type "date", since IMAP
servers like this not so common date format.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}

\item{by}{Which id (\code{MSN} or \code{UID}) to use when searching for
messages. \code{"MSN"} or message squence number is a message's relative
position to the older message in the mailbox. It may change after deleting
or moving messages. For instance, if a message is deleted, message sequence
numbers are reordered to fill the gap. \code{"UID"} or unique identifier
is always the same during the life cycle of a message. Default is
\code{"MSN"}.}

\item{flag}{Optional parameter that adds a flag filter to the search. Use
\link{flag_options} to list the common flags used by IMAP servers.
Default is \code{NULL}.}

\item{esearch}{A logical. Default is \code{FALSE}. If the IMAP server has
\code{ESEARCH} capability, it can be used to optimize search results. It
allows to condense results to message sets to cut down on transmission
costs, e.g. \code{1:5} instead of writing ids individually such as
code{1,2,3,4,5}. It can be used along with buffersize to avoid results
stripping. Check if your IMAP server supports \code{SEARCH} with
\code{\link{list_server_capabilities}}.}

\item{return_imapconf}{A logical. If \code{TRUE}, the function returns a
\code{list} of length \code{2}, containing the \code{imapconf} object
(IMAP settings) and the search results as message ids (\code{"MSN"} or
\code{"UID"}). If \code{FALSE}, returns only the message ids as a numeric
vector. Default is \code{TRUE}.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
Depending on the \code{return_imapconf} parameter, it can be a
    \code{list} of length \code{2} containing the \code{imapconf} object
    and message ids resulted from search, or a numeric vector containing
    only message ids.
}
\description{
Functions that allows searching for messages using the (internal)
    date criteria, such as before, since, on, and period.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configure_imap(url="imaps://export.imap.aol.com",
                           username="your_AOL_user",
                           password=rstudioapi::askForPassword()
                          )

# search 1
result1 <- imapconf \%>\%
    select_mailbox(mbox = "Sent") \%>\%
    search_since(date_char = "17-Apr-2012")

# search 2
result2 <- imapconf \%>\%
    select_mailbox(mbox = "Inbox") \%>\%
    search_since(date_char = "17-Jun-2019", flag = "FLAGGED")
# depending on results, it might be necessary to increase buffersize in
# configure_imap()

}

}
\references{
Resnick, P., Ed., "Internet Message Format", RFC 5322,
    \doi{10.17487/RFC5322}, October 2008, \url{https://www.rfc-editor.org/info/rfc5322}.

Babcock, N., "Introduction to IMAP", Blog, May 2016,
    \url{https://nbsoftsolutions.com/blog/introduction-to-imap}.
}
\seealso{
Other Date-search operations: \code{\link{custom_search}},
  \code{\link{search_before}}, \code{\link{search_on}},
  \code{\link{search_period}},
  \code{\link{search_sent_before}},
  \code{\link{search_sent_on}},
  \code{\link{search_sent_period}},
  \code{\link{search_sent_since}}
}
\concept{Date-search operations}
