% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMailbox.R
\name{selectMailbox}
\alias{selectMailbox}
\title{Mailbox Selection}
\usage{
selectMailbox(imapconf, mbox, retries = 2)
}
\arguments{
\item{imapconf}{A object of class \code{imapconf} generated with
\link{configureIMAP}.}

\item{mbox}{A string containing the name of an existing mailbox on the server.}

\item{retries}{Number of attempts to connect and execute the command. Default
is \code{2}.}
}
\value{
The \code{imaconf, list} object with the selected mailbox added to it.
}
\description{
Select a mailbox on the server for issuing further commands.
}
\examples{
\dontrun{

# configure IMAP
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# Select INBOX
imapconf \%>\%
  selectMailbox(mbox = "INBOX")

}
}
\seealso{
Other mailbox commands: \code{\link{examineMailbox}},
  \code{\link{listMailboxes}},
  \code{\link{listServerCapabilities}},
  \code{\link{renameMailbox}}
}
\concept{mailbox commands}
