% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaller-than.R
\name{smaller_than}
\alias{smaller_than}
\title{Criterion constructor function to be combined in a custom search statement}
\usage{
smaller_than(size, negate = FALSE)
}
\arguments{
\item{size}{An integer specifying the number of bytes to be used as
search criterion.}

\item{negate}{If \code{TRUE}, negates the search and seeks for "NOT SEARCH
CRITERIA". Default is \code{FALSE}.}
}
\description{
Criterion constructor function to be combined in a custom search statement
}
\examples{
\dontrun{
# select folder & search
con$select_folder(name = "INBOX")
# search for messages containing the string "XYZ@k-state.edu" in the
# "FROM" field OR those that are SMALLER than 512KB.
res <- con$search(request = OR(string(expr = "XYZ@k-state.edu",
                                      where = "FROM"),
                               smaller_than(size = 512000)))
}

}
\seealso{
Other custom search:
\code{\link[=AND]{AND()}},
\code{\link{ImapCon}},
\code{\link[=OR]{OR()}},
\code{\link[=before]{before()}},
\code{\link[=flag]{flag()}},
\code{\link[=larger_than]{larger_than()}},
\code{\link[=modseq]{modseq()}},
\code{\link[=older_than]{older_than()}},
\code{\link[=on]{on()}},
\code{\link[=saved_before]{saved_before()}},
\code{\link[=sent_before]{sent_before()}},
\code{\link[=sent_on]{sent_on()}},
\code{\link[=sent_since]{sent_since()}},
\code{\link[=since]{since()}},
\code{\link[=string]{string()}},
\code{\link[=younger_than]{younger_than()}}
}
\concept{custom search}
