% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallerThan.R
\name{smallerThan}
\alias{smallerThan}
\title{Criteria Helper Functions for Custom Search}
\usage{
smallerThan(size, negate = FALSE)
}
\arguments{
\item{size}{An integer specifying the size (in number characters) of message to be used as search
criterion.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Criteria helper functions to be combined inside a operator helper
    function as a custom request in \link{customSearch}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request =
                 OR(String(
                       section_or_field = "from", string = "allanvcq@gmail.com"),
                    smallerThan(
                       size = 512000)
                   )
                 )
# searches for messages containing the string "allanvcq@gmail.com" in the field "FROM"
    OR those smallerThan 512KB.

}

}
\seealso{
Other customsearch helper functions: \code{\link{AND}},
  \code{\link{Before}}, \code{\link{Flag}},
  \code{\link{OR}}, \code{\link{Since}},
  \code{\link{String}}, \code{\link{olderThan}},
  \code{\link{sentBefore}}, \code{\link{sentOn}},
  \code{\link{sentSince}}, \code{\link{youngerThan}}
}
\concept{customsearch helper functions}
