% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{string}
\alias{string}
\title{Criterion constructor function to be combined in a custom search statement}
\usage{
string(expr, where, negate = FALSE)
}
\arguments{
\item{expr}{A character string specifying the word or expression to search
for in messages.}

\item{where}{A mandatory character string specifying in which
message's Section or Header Field to search for the provided string.}

\item{negate}{If \code{TRUE}, negates the search and seeks for "NOT SEARCH
CRITERIA". Default is \code{FALSE}.}
}
\description{
Criterion constructor function to be combined in a custom search statement
}
\examples{
\dontrun{
# select folder & search
con$select_folder(name = "INBOX")
# search for messages containing the string "XYZ@k-state.edu" in the
#   "FROM" AND the string "@gmail.com" in the "CC" field.
res <- con$search(request = AND(string(expr = "XYZ@k-state.edu",
                                      where = "FROM"),
                               string(expr = "@gmail.com",
                                      where = "CC")))
}

}
\seealso{
Other custom search:
\code{\link[=AND]{AND()}},
\code{\link{ImapCon}},
\code{\link[=OR]{OR()}},
\code{\link[=before]{before()}},
\code{\link[=filter_stored]{filter_stored()}},
\code{\link[=flag]{flag()}},
\code{\link[=fuzzy]{fuzzy()}},
\code{\link[=imap_raw]{imap_raw()}},
\code{\link[=larger_than]{larger_than()}},
\code{\link[=modseq]{modseq()}},
\code{\link[=older_than]{older_than()}},
\code{\link[=on]{on()}},
\code{\link[=saved_before]{saved_before()}},
\code{\link[=sent_before]{sent_before()}},
\code{\link[=sent_on]{sent_on()}},
\code{\link[=sent_since]{sent_since()}},
\code{\link[=since]{since()}},
\code{\link[=smaller_than]{smaller_than()}},
\code{\link[=younger_than]{younger_than()}}
}
\concept{custom search}
