% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AND.R
\name{AND}
\alias{AND}
\title{Relational Operators - Helper Functions for Custom Search}
\usage{
AND(..., negate = FALSE)
}
\arguments{
\item{...}{a combination of criteria helper functions with its arguments.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Allows a combination of criteria helper
    functions such as \link{Before}, \link{Since}, \link{On},
    \link{sentBefore}, \link{sentSince}, \link{sentOn}, \link{Flag}, or
    \link{String}, in order to execute a multiple criteria custom search.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request = AND(sentSince(date.char = "17-Apr-2019"),
                                      smallerThan(size = 512000),
                                      negate = TRUE))
# searches for messages NOT SentSince "17-Apr-2019" AND NOT SmallerThan  512KB.

}

}
\seealso{
Other customsearch helper functions: \code{\link{Before}},
  \code{\link{Flag}}, \code{\link{OR}},
  \code{\link{Since}}, \code{\link{String}},
  \code{\link{olderThan}}, \code{\link{sentBefore}},
  \code{\link{sentOn}}, \code{\link{sentSince}},
  \code{\link{smallerThan}}, \code{\link{youngerThan}}
}
\concept{customsearch helper functions}
