% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag}
\alias{Flag}
\title{Criteria Helper Functions for Custom Search}
\usage{
Flag(flag, negate = FALSE)
}
\arguments{
\item{flag}{A string specifying the flag to be used for filtering messages.
Use \link{flag_options} to list the flags available in your IMAP server.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Criteria helper functions to be combined inside a operator helper
    function as a custom request in \link{customSearch}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request = AND(Flag(flag="UNSEEN"),
                                      smallerThan(size = 512000),
                                      negate = TRUE))
# searches for messages with Flag ""UNSEEN" AND NOT SmallerThan  512KB.

}

}
\seealso{
Other customsearch helper functions: \code{\link{AND}},
  \code{\link{Before}}, \code{\link{OR}},
  \code{\link{Since}}, \code{\link{String}},
  \code{\link{olderThan}}, \code{\link{sentBefore}},
  \code{\link{sentOn}}, \code{\link{sentSince}},
  \code{\link{smallerThan}}, \code{\link{youngerThan}}
}
\concept{customsearch helper functions}
