% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-options.R
\name{metadata_options}
\alias{metadata_options}
\title{Message Metadata Options}
\usage{
metadata_options()
}
\value{
A \code{vector} containing message metadata fields.
}
\description{
List Metadata fields used in messages.
}
\note{
This function lists message metadata used by
    IMAP servers, according to the RFC 2060 (Crispin, 1996).
}
\examples{
\dontrun{

library(mRpostman)
metadata_options()

}
}
\references{
Crispin, M., "Internet Message Access Protocol - Version 4rev1",
    RFC 2060, \doi{10.17487/RFC2060}, December 1996,
    \url{https://www.rfc-editor.org/info/rfc2060}.
}
\seealso{
Other options: \code{\link{flag_options}},
  \code{\link{section_or_field_options}}
}
\concept{options}
