\docType{package}
\name{mRpostman-package}
\alias{mRpostman-package}
\alias{mRpostman}
\title{An IMAP client for R}

\description{
  \pkg{mRpostman} is an easy-to-use IMAP client that provides tools for searching and message fetching, mailbox management, attachment extraction, and several other IMAP features, hence paving the way for email data analysis from within \R. To do so, this package makes extensive use of the \pkg{curl} package and the libcurl C library..
}


\author{

  Author & Mantainer: Allan Quadros \email{allanvcq@gmail.com}

}

\references{

Crispin, M. (2003), \emph{INTERNET MESSAGE ACCESS PROTOCOL - VERSION 4rev1}, RFC 3501, March 2003, \url{https://tools.ietf.org/html/rfc3501}.

Heinlein, P. and Hartleben, P. (2008). \emph{The Book of IMAP: Building a Mail Server with Courier and Cyrus}. No Starch Press. ISBN 978-1-59327-177-0.

Ooms, J. (2020). \emph{curl: A Modern and Flexible Web Client for R}. R package version 4.3, \url{https://CRAN.R-project.org/package=curl}.

Stenberg, D. \emph{Libcurl - The Multiprotocol File Transfer Library}, \url{https://curl.haxx.se/libcurl/}.

}

\seealso{

  Useful links:

  \itemize{
    \item{ \code{mRpostman official website}: \url{https://allanvc.github.io/mRpostman}}
  }

}
