\docType{package}
\name{mRpostman-package}
\alias{mRpostman-package}
\alias{mRpostman}
\title{IMAP Tools for R}

\description{
  \pkg{mRpostman} provides multiple IMAP (Internet Message Access Protocol) 
  commands based on the RFC 3501 manual (Crispin, 2003), its updates, and other 
  related documents. The goal is to make it easy to connect to your IMAP server 
  and execute commands such as listing mailboxes, searching and fetching 
  messages in a tidy way.
}


\author{

  Author & Mantainer: Allan Quadros \email{allanvcq@gmail.com}

}

\references{

Crispin, M., \emph{INTERNET MESSAGE ACCESS PROTOCOL - VERSION 4rev1}, RFC 3501, 
DOI: 10.17487/RFC3501, March 2003, \url{https://www.rfc-editor.org/info/rfc3501}.

Ooms, J. \emph{curl: A Modern and Flexible Web Client for R}. R package version 3.3, 
2019, \url{https://CRAN.R-project.org/package=curl}.

Stenberg, D. \emph{Libcurl - The Multiprotocol File Transfer Library}, 
\url{https://curl.haxx.se/libcurl/}.

}

\seealso{

  Useful links:
  
  \itemize{
    \item{ \code{mRpostman official website}: \url{https://allanvc.github.io/mRpostman}}
  }

}
