% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/On.R
\name{On}
\alias{On}
\title{Criteria Helper Functions for Custom Search}
\usage{
On(date_char, negate = FALSE)
}
\arguments{
\item{date_char}{A character vector with format "DD-Mon-YYYY",
e.g. "01-Apr-2019". We opted not to use objects of type "date", since IMAP
servers like this not so common date format.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Criteria helper functions to be combined inside a operator helper
    function as a custom request in \link{customSearch}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request = OR(On(date_char = "17-Apr-2019"),
                                     On(date_char = "17-Jul-2018")))
# searches for messages On "17-Apr-2019" OR On "17-Jul-2018".

}

}
\seealso{
Other search: \code{\link{largerThan}}

Other custom: \code{\link{largerThan}}

Other criteria helper functions: \code{\link{largerThan}}
}
\concept{criteria helper functions}
\concept{custom}
\concept{search}
