% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent-before.R
\name{sent_before}
\alias{sent_before}
\title{Criterion constructor function to be combined in a custom search statement}
\usage{
sent_before(date_char, negate = FALSE)
}
\arguments{
\item{date_char}{A \code{character string} with format "DD-Mon-YYYY", e.g.
"01-Apr-2019". We opt not to use \code{Date} or \code{POSIX*} like
objects, since IMAP servers use this unusual date format.}

\item{negate}{If \code{TRUE}, negates the search and seeks for "NOT SEARCH
CRITERIA". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{request} parameter in
 \code{ImapCon$search()} function.
}
\description{
Criterion constructor function to be combined in a custom search statement
}
\examples{
\dontrun{
# select folder & search
con$select_folder(name = "INBOX")
# search for messages SINCE "30-Aug-2019" AND SMALLER than 512KB.
res <- con$search(request = AND(sent_since(date_char = "30-Aug-2019"),
                                smaller_than(size = 512000)))
}
}
\seealso{
Other custom search:
\code{\link[=AND]{AND()}},
\code{\link{ImapCon}},
\code{\link[=OR]{OR()}},
\code{\link[=before]{before()}},
\code{\link[=flag]{flag()}},
\code{\link[=larger_than]{larger_than()}},
\code{\link[=older_than]{older_than()}},
\code{\link[=on]{on()}},
\code{\link[=sent_on]{sent_on()}},
\code{\link[=sent_since]{sent_since()}},
\code{\link[=since]{since()}},
\code{\link[=smaller_than]{smaller_than()}},
\code{\link[=string]{string()}},
\code{\link[=younger_than]{younger_than()}}
}
\concept{custom search}
