% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/String.R
\name{String}
\alias{String}
\title{Criteria Helper Functions for Custom Search}
\usage{
String(section_or_field, string, negate = FALSE)
}
\arguments{
\item{section_or_field}{A mandatory character string specifying in which
messages's Section or Header Field to search for the provided string.
For some available options, see \link{section_or_field_options}.}

\item{string}{A character string specifying the word or expression to
search for in messages.}

\item{negate}{If \code{TRUE}, negates the search and seeks for
"NOT search_criterion". Default is \code{FALSE}.}
}
\value{
A search string to be used as a \code{custom_request} parameter in
    \link{customSearch} function.
}
\description{
Criteria helper functions to be combined inside a operator helper
    function as a custom request in \link{customSearch}.
}
\examples{

\dontrun{

# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf \%>\%
    selectMailbox(mbox = "INBOX") \%>\%
    customSearch(custom_request =
                 OR(String(
                       section_or_field = "from", string = "allanvcq@gmail.com"),
                    String(
                       section_or_field = "from", string = "allanvcq@yahoo.com")
                   )
                 )
# searches for messages containing the string "allanvcq@gmail.com" OR
    "allanvcq@yahoo.com" in the "FROM" field.

}

}
\seealso{
Other customsearch helper functions: \code{\link{AND}},
  \code{\link{Before}}, \code{\link{Flag}},
  \code{\link{OR}}, \code{\link{Since}},
  \code{\link{olderThan}}, \code{\link{sentBefore}},
  \code{\link{sentOn}}, \code{\link{sentSince}},
  \code{\link{smallerThan}}, \code{\link{youngerThan}}
}
\concept{customsearch helper functions}
